requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
jwt
joblib
requests
